Refactor state management to Redux, fix initial loading and request sending
